\name{catts1}
\docType{data}
\alias{catts1}
    
\title{Categorical eye-movement series (listener)}

\description{
  
  An eye-movement fixation scan-pattern, i.e.,
  a temporal sequence of fixated objects.
  
}

\usage{catts1}
\format{A vector containing 605 observations.}

\references{

  Coco, M., Dale, R., and Keller, F. (n.d.). 
  The role of interactivity on cognitive alignment 
  and decision making during dialogue. (submitted)
  
}

\keyword{datasets}