\name{contts1}
\docType{data}
\alias{contts1}
    
\title{Continuous eye-movement series (listener)}

\description{
  
  Eye-movement sequence of visual saliency value
  observed at fixation location.
  
}

\usage{contts1}
\format{A vector containing 605 observations.}

\references{

  Coco, M., Dale, R., and Keller, F. (n.d.). 
  The role of interactivity on cognitive alignment 
  and decision making during dialogue. (submitted)
  
}

\keyword{datasets}