\name{eyemovement}
\docType{data}
\alias{eyemovement}
    
\title{Eye-movement categorical time-series}

\description{
  
  A two-columns dataset of eye-movement fixation scan-pattern, 
  which are temporal sequences of fixated objects.

}

\usage{eyemovement}
\format{A data frame with 1000 rows and 2 variables:
 \describe{
   \item{listener}{the listener time series}
   \item{narrator}{the narrator time series}
  }
 }


\references{

  Richardson, D. C., and Dale, R. (2005). Looking to
  understand: The coupling between speakers and listeners
  eye movements and its relationship to discourse
  comprehension. Cognitive Science, 29, 39-54.

}

\keyword{datasets}