\name{Figure_1}
\docType{data}
\alias{Figure_1}
    
\title{Eye-movement categorical time-series}

\description{
  
  Three time series (1250 observations) organised as columns and simulating data from: a periodical sine wave, one of the dimensions of the Lorenz attractor and a white noise signal. 

}

\usage{Figure_1}
\format{A data frame with 1250 rows and 3 columns:
 \describe{
   \item{sinus}{A periodical sine wave}
   \item{lorenz}{One dimension of a Lorenz attractor}
   \item{wnoise}{White noise}

  }
 }

\keyword{datasets}