\name{Figure_2}
\docType{data}
\alias{Figure_2}
    
\title{A unidimensional sinusoidal time series}

\description{
  
  A unidimensional sinusoidal time series

}

\usage{Figure_2}
\format{A data frame with 45 rows and 1 column:
 \describe{
   \item{V1}{A unidimensional sinusoidal time series}
  }
 }


\keyword{datasets}