\name{Figure_3}
\docType{data}
\alias{Figure_3}
    
\title{Simulated time series of the three dimensions from the Lorenz system}

\description{
  
   Simulated time series of the three dimensions from the Lorenz system

}

\usage{Figure_3}
\format{A data frame with 2048 rows and 3 columns:
 \describe{
   \item{V1}{First dimension of the Lorenz system}
   \item{V2}{Second dimension of the Lorenz system}
   \item{V3}{Third dimension of the Lorenz system}
  }
 }


\keyword{datasets}