\name{handmovement}
\docType{data}
\alias{handmovement}
    
\title{Continuous series of hand movements}

\description{	

Hand-movement velocity profiles of two participants (P1 and P2) for the dominant (\_d) and non-dominant (\_n) hand. 

}

\usage{handmovement}

\format{A dataframe of 5799 observations.

\describe{
  \item{P1_TT_d}{Participant 1 dominant hand}
  \item{P1_TT_n}{Participant 1 non-dominant hand}
  \item{P2_TT_d}{Participant 2 dominant hand}
  \item{P2_TT_n}{Participant 2 non-dominant}
 }

}

\references{

Wallot, S., Mitkidis, P., McGraw, J. J. and Roepstorff, A. (2016). Beyond synchrony: joint action in a complex production task reveals beneficial effects of decreased interpersonal synchrony. PloS one, 11(12), e0168306.

}

\keyword{datasets}
