\name{RDts1}
\docType{data}
\alias{RDts1}
    
\title{Categorical eye-movement series (listener)}

\description{
  
  An eye-movement fixation scan-pattern, i.e.,
  a temporal sequence of fixated objects.
  
}

\usage{RDts1}
\format{A vector containing 2000 observations.}

\references{

  Richardson, D. C., and Dale, R. (2005). Looking to
  understand: The coupling between speakers and listeners
  eye movements and its relationship to discourse
  comprehension. Cognitive Science, 29, 39-54.

}

\keyword{datasets}