\name{rightmov}
\docType{data}
\alias{rightmov}
    
\title{Continuous series of body movement intensity (left participant)}

\description{	
  	A sequence of z-score gross body movement,
        as captured in a video, sampled at 8 Hz  
}

\usage{rightmov}
\format{A vector containing 801 observations.}

\references{
  Paxton, A., and Dale, R. (2013). Frame-differencing methods 
  for measuring bodily synchrony in conversation. 
  Behavior research methods, 45(2), 329-343. 
}

\keyword{datasets}
