\name{text}
\docType{data}
\alias{text}
    
\title{Categorical sequence of words}

\description{

	A simple text of the nursery rhyme, \'the wheels on the bus\`   
  
}


\usage{text}
\format{A vector of 120 words}

\references{

Verna Hills (1939). The Wheels on the bus. American Childhood (25), 56

}

\keyword{datasets}
